Exrtract 'useResetParameterMapping'
